requirements.txt
astroid==2.8.6
bidict==0.21.2
blinker==1.4
cachelib==0.4.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
dnspython==1.16.0
email-validator==1.1.2
eventlet==0.30.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Session==0.4.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
isort==5.10.1
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.4
platformdirs==2.4.0
postgres==4.0
psycopg2==2.8.6
psycopg2-binary==2.9.2
psycopg2-pool==1.1
pycparser==2.20
PyJWT==2.0.1
pylint==2.11.1
python-dotenv==0.15.0
python-engineio==4.0.1
python-socketio==5.1.0
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.7
tmdbsimple==2.8.0
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.13.3
WTForms==2.3.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0
